% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function to a data frame}
\usage{
df_apply(
  .data,
  .f,
  .condition = function(x) {     TRUE },
  .else = function(x) {     x },
  ...
)
}
\arguments{
\item{.data}{A data frame}

\item{.f}{A function you would like to apply to a column(s) of the data frame}

\item{.condition}{A condition that column must have to apply the function}

\item{.else}{If the condition fails, a second function to apply}

\item{...}{Allows for arguments to be passed into .f}
}
\value{
A data frame
}
\description{
Apply a function to a data frame
}
\examples{
n <- 7L
TestData <-
 tibble( double = rnorm(n, 100, 10),
         x = 123400 / 10^(1L:n),
         integer = (1L:n) * (1L:n),    # ^2 would return a double!!
         character = LETTERS[1L:n],
         factor = factor(letters[1L:n]),
         logical = rep(c(TRUE, FALSE), length.out = n) )
df_apply(TestData, round, is.numeric, toupper, digits = 1)

load(medals_per_event)
df_apply(medals_per_event, tolower, !is.numeric)
}
