% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympic-datasets.R
\docType{data}
\name{medal_counts}
\alias{medal_counts}
\title{Medal Counts for Winter Olympics by Country

A dataset containing the medals and places won by each country, assuming the country
has won at least 100 total medals}
\format{
A data frame with 42 rows and 3 variables:
\describe{
\item{Country}{name of country, as referred to by the USA}
\item{Medal}{type of medal, factor with levels "gold", "silver" or "bronze"}
\item{totals}{number of medals of that type, in counts}
}
}
\usage{
medal_counts
}
\description{
Medal Counts for Winter Olympics by Country

A dataset containing the medals and places won by each country, assuming the country
has won at least 100 total medals
}
\examples{
head(medal_counts)

medal_counts |>
mutate(Medal = forcats::fct_relevel(Medal,'bronze','silver','gold'),
      Country = reorder(Country, totals, sum)) |>
 plot_ly(x = ~Country, y = ~totals, color = ~Medal) \%>\%
 add_bars() |>
 layout(barmode = "stack",
        hovermode = "x") |>
 layout(title = "Norway Has the Most Gold, Silver,
 <br> and Bronze Winter Olympic Medals <br> (1924 - 2014)")
}
\keyword{datasets}
