% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympic-datasets.R
\docType{data}
\name{medal_location2}
\alias{medal_location2}
\title{Medal Counts for Winter Olympics by Country

A dataset containing the medals and places won by each country, assuming the country
has won at least 1 total medal; helpful for geospatial mapping}
\format{
A data frame with 36 rows and 6 variables:
\describe{
\item{name}{name of country, as referred to by the USA}
\item{iso_a3}{country code, as referred to by the UN}
\item{bronze}{number of bronze medals, in counts}
\item{gold}{number of gold medals, in counts}
\item{silver}{number of silver medals, in counts}
\item{total}{total number of medals, sum of bronze, gold, and silver}
}
}
\usage{
medal_location2
}
\description{
Medal Counts for Winter Olympics by Country

A dataset containing the medals and places won by each country, assuming the country
has won at least 1 total medal; helpful for geospatial mapping
}
\examples{
head(medal_location2)

plot_ly(medal_location2,
type='choropleth',
locations=medal_location2$iso_a3,
z=medal_location2$total, text=medal_location2$hover, colorscale="Blues",
reversescale = TRUE) |> # make darker countries = more medals
 layout(title = "Winter Olympic Medals Won per Country <br> (1924 - 2014)") |>
 colorbar(title = "Total Medals")
}
\keyword{datasets}
