% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympic-datasets.R
\docType{data}
\name{medals_per_event}
\alias{medals_per_event}
\title{Medal Counts for Winter Olympics by Country per event

A dataset containing the medals and places won by each country,
in each event, assuming the country has won at least 1 total medal}
\format{
A data frame with 565 rows and 4 variables:
\describe{
\item{Country}{name of country, as referred to by the USA}
\item{Medal}{type of medal, factor with levels "gold", "silver" or "bronze"}
\item{Sport}{name of sport, as recognized by the IOC}
\item{totals}{number of medals of that type, in counts}
}
}
\usage{
medals_per_event
}
\description{
Medal Counts for Winter Olympics by Country per event

A dataset containing the medals and places won by each country,
in each event, assuming the country has won at least 1 total medal
}
\examples{
glimpse(medals_per_event)

bscols(widths = c(2,NA),
list(
 filter_select("Sport", "Sport", shared_medals, ~Sport)
),
ggplotly(shared_medals \%>\%
          ggplot(aes(Country, totals)) +
          geom_point(aes(color = Medal, label = Sport))  +
          theme(axis.text.x = element_blank(), axis.ticks.x = element_blank()) +
          scale_color_manual(values=c("#FFD700", "#C0C0C0", "#CD7F32")) +
          labs(x = "Country <br> (alphabetical)", y = "Medals Won", size = "",
               title = "Medals per Country <br> Split by Event")))
}
\keyword{datasets}
