% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympic-datasets.R
\docType{data}
\name{total_medals}
\alias{total_medals}
\title{Total Winter Olympic Medals won by each country

A dataset containing the total medals won by each country, assuming the country
has won at least 100 medals}
\format{
A data frame with 14 rows and 2 variables:
\describe{
\item{Country}{name of country, as referred to by the USA}
\item{totals}{total number of medals, in counts}
}
}
\usage{
total_medals
}
\description{
Total Winter Olympic Medals won by each country

A dataset containing the total medals won by each country, assuming the country
has won at least 100 medals
}
\examples{
pillar::glimpse(total_medals)

total_medals |>
plotly::plot_ly(x = ~Country, y = ~totals) |>
 plotly::add_bars() |>
 plotly::layout(title = "Norway has more Winter Olympic
 Medals than All Other Countries<br> (1924 - 2014)")
}
\keyword{datasets}
