% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function to a data frame}
\usage{
df_apply(
  .data,
  .f,
  .condition = function(x) {     TRUE },
  .else = function(x) {     x },
  ...
)
}
\arguments{
\item{.data}{A data frame}

\item{.f}{A function you would like to apply to a column(s) of the data frame}

\item{.condition}{A condition that column must have to apply the function}

\item{.else}{If the condition fails, a second function to apply}

\item{...}{Allows for arguments to be passed into .f}
}
\value{
A data fram
}
\description{
Apply a function to a data frame
}
